requirements.txt
Django==4.2.10
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.2.0
django-allauth==0.61.1
django-environ==0.11.2
django-storages==1.14.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
celery==5.3.6
redis==5.0.1
stripe==7.11.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-cleanup==8.0.0
django-ckeditor==6.7.0
django-redis==5.4.0
django-cacheops==7.2
django-ratelimit==4.1.0
django-simple-history==3.4.0
django-import-export==3.3.4
django-admin-honeypot==1.1.0
django-axes==6.3.0
django-otp==1.3.0
django-two-factor-auth==1.15.5
django-recaptcha==4.0.0
django-newsletter==0.9.0
channels==4.0.0
daphne==4.0.0 